Fixing typo in README
Thanks for sharing, this was very helpful! I noticed a typo in the README, so I wanted to give back a little by fixing it. Also, I noticed that the latest image on GCR wasn't tagged as `latest` so some things weren't working right. Adding the `v2.3.0` tag fixed things up!